pyproject.toml
[project]
name = "icc-eval-etl"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "httpx>=0.28",
    "pydantic>=2.10",
    "typer>=0.15",
    "pyyaml>=6.0",
]
